fix exception during UglifyJS.minify()
During a `roots compile -e production`, the accumulated javascript was getting
a syntax error between two files in the pipeline, which threw an exception
during the UglifyJS.minify() step. Specifically, the `all_contents` var
contained this bit of javascript:

    .... [tail end of first javascript file]
    prototype.toString=function(){return""+this._wrapped},"function"==typeof
    define&&define.amd&&define("underscore",[],function(){return m})}).call(this);
    //# sourceMappingURL=underscore-min.map(function() {

        var a, b, c;
    ....
    }).call(this);
    .... [tail end of next javascript file]

Note that the tail end of the first file has the `sourceMappingURL` in a
comment, and the next file gets concat'd immediately after. However, the start
of the second file gets lost in the commented-out `sourceMappingURL` of the
first file. The .js should look more like:

    .... [tail end of first javascript file]
    prototype.toString=function(){return""+this._wrapped},"function"==typeof
    define&&define.amd&&define("underscore",[],function(){return m})}).call(this);
    //# sourceMappingURL=underscore-min.map
    (function() {

        var a, b, c;
    ....
    }).call(this);
    .... // tail end of next javascript file